backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.2.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.5
langchain==0.1.0
langchain-openai==0.0.5
requests==2.31.0
firebase-admin==6.2.0
cors==1.0.1
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.23
